refactor : redux-actions를 적용해 if문 없애기 적용
