+++ package.json
@@ -21,6 +21,7 @@
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.7",
     "@types/jest": "^26.0.23",
+    "@types/redux-actions": "^2.6.2",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
     "codeceptjs": "^3.0.7",
@@ -42,6 +43,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^8.0.5",
-    "redux": "^4.2.1"
+    "redux": "^4.2.1",
+    "redux-actions": "^3.0.0"
   }
 }
