+++ src/InputContainer.jsx
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 import Input from './Input';
-import { updateTaskTitle, addTask } from './actions';
+import { updateTaskTitle, addTask } from './reducers';
 
 export default function InputContainer() {
   const { taskTitle } = useSelector((state) => ({
@@ -9,11 +9,11 @@ export default function InputContainer() {
   const dispatch = useDispatch();
   // 관심사분리
   function handleChangeTitle(event) {
-    dispatch(updateTaskTitle(event.target.value));
+    dispatch(updateTaskTitle({ taskTitle: event.target.value }));
   }
 
   function handleClickAddTask() {
-    dispatch((addTask()));
+    dispatch(addTask());
   }
 
   return (
