--- src/reducer.js
@@ -1,43 +1,0 @@
-// Redux action
-// - type(string)
-// - payload(object)
-
-const initialState = {
-  newId: 100,
-  taskTitle: '',
-  tasks: [],
-};
-
-export default function reducer(state = initialState, action) {
-  if (action.type === 'updateTaskTitle') {
-    return {
-      ...state,
-      taskTitle: action.payload.taskTitle,
-    };
-  }
-
-  if (action.type === 'addTask') {
-    const { newId, taskTitle, tasks } = state;
-
-    if (taskTitle === '') {
-      return state;
-    }
-    return {
-      ...state,
-      newId: newId + 1,
-      taskTitle: '',
-      tasks: [...tasks, { id: newId, title: taskTitle }],
-    };
-  }
-
-  if (action.type === 'deleteTask') {
-    const { tasks } = state;
-    return {
-      ...state,
-      tasks: tasks.filter((task) => task.id !== action.payload.id),
-
-    };
-  }
-
-  return state;
-}
