+++ src/store.js
@@ -1,7 +1,7 @@
 import { createStore } from 'redux';
 
-import reducer from './reducer';
+import reducer from './reducers';
 
 const store = createStore(reducer);
-
+console.log(store);
 export default store;
